+++ scripts/test-tenant-commands.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Copyright (c) 2026 eLeDia.de / Bernd Schreistetter (bsc)
-# Version: v3.4.11
+# Version: v3.4.12
 #
 # eLeDia tenant command matrix test.
 # Exercises solr-tenant.sh commands against a running Solr container.